robustness fixes
