+++ environment/commands/transfer-eth-erc20-from-near.js
@@ -9,7 +9,11 @@ const { NearMintableToken } = require('../lib/near-mintable-token')
 const { RainbowConfig } = require('../lib/config')
 const { borshifyOutcomeProof } = require('../lib/borsh')
 const { sleep, RobustWeb3 } = require('../lib/robust')
-const { normalizeEthKey } = require('../lib/robust')
+const {
+  normalizeEthKey,
+  backoff,
+  nearJsonContractFunctionCall,
+} = require('../lib/robust')
 
 class TransferEthERC20FromNear {
   static async execute(command) {
@@ -47,9 +51,11 @@ class TransferEthERC20FromNear {
     await nearTokenContractBorsh.accessKeyInit()
 
     // Burn the token on Near side.
-    const old_balance = await nearTokenContract.get_balance({
-      owner_id: command.nearSenderAccount,
-    })
+    const old_balance = await backoff(10, () =>
+      nearTokenContract.get_balance({
+        owner_id: command.nearSenderAccount,
+      })
+    )
     console.log(
       `Balance of ${command.nearSenderAccount} before burning: ${old_balance}`
     )
@@ -59,7 +65,8 @@ class TransferEthERC20FromNear {
     console.log(
       `Burning ${command.amount} tokens on NEAR blockchain in favor of ${ethReceiverAddress}.`
     )
-    const txBurn = await nearTokenContractBorsh.functionCall(
+    const txBurn = await nearJsonContractFunctionCall(
+      RainbowConfig.getParam('near-fun-token-account'),
       nearSenderAccount,
       'burn',
       { amount: command.amount, recipient: ethReceiverAddress },
@@ -188,9 +195,11 @@ class TransferEthERC20FromNear {
     }
 
     console.log(`Burnt ${JSON.stringify(command.amount)}`)
-    const new_balance = await nearTokenContract.get_balance({
-      owner_id: command.nearSenderAccount,
-    })
+    const new_balance = await backoff(10, () =>
+      nearTokenContract.get_balance({
+        owner_id: command.nearSenderAccount,
+      })
+    )
     console.log(
       `Balance of ${command.nearSenderAccount} after burning: ${new_balance}`
     )
