+++ environment/lib/robust.js
@@ -2,6 +2,7 @@
 /// and wrap web3 with error handling and retry
 const Web3 = require('web3')
 const _ = require('lodash')
+const nearlib = require('near-api-js')
 
 const RETRY = 10
 const DELAY = 500
@@ -12,6 +13,7 @@ const retry = (retries, fn) =>
     retries > 1 ? retry(retries - 1, fn) : Promise.reject(err)
   )
 const sleep = duration => new Promise(res => setTimeout(res, duration))
+
 const backoff = (retries, fn, delay = DELAY, wait = BACKOFF) =>
   fn().catch(err =>
     retries > 1
@@ -172,6 +174,128 @@ const promiseWithTimeout = (timeoutMs, promise, failureMessage) => {
   })
 }
 
+async function nearJsonContractFunctionCall(
+  contractId,
+  sender,
+  method,
+  args,
+  gas,
+  amount
+) {
+  // A robust version of near-api-js account.functionCall. We can't simply retry account.functionCall because
+  // we don't know whether txn successfully submitted when timeout, so there's a risk of double sending
+
+  await sender.ready
+  let accessKey = await sender.findAccessKey()
+  return await signAndSendTransaction(accessKey, sender, contractId, [
+    nearlib.transactions.functionCall(
+      method,
+      Buffer.from(JSON.stringify(args)),
+      gas,
+      amount
+    ),
+  ])
+}
+
+const RETRY_SEND_TX = 10
+const RETRY_TX_STATUS = 10
+
+const signAndSendTransaction = async (
+  accessKey,
+  account,
+  receiverId,
+  actions
+) => {
+  // TODO: Find matching access key based on transaction
+  let errorMsg
+  let resendLast = false
+  let sendTxnAsync
+  let txHash
+
+  for (let i = 0; i < RETRY_SEND_TX; i++) {
+    try {
+      if (resendLast) {
+        console.log('resend txn')
+        await sendTxnAsync()
+        resendLast = false
+      } else {
+        const status = await account.connection.provider.status()
+        let signedTx
+        ;[txHash, signedTx] = await nearlib.transactions.signTransaction(
+          receiverId,
+          ++accessKey.nonce,
+          actions,
+          nearlib.utils.serialize.base_decode(
+            status.sync_info.latest_block_hash
+          ),
+          account.connection.signer,
+          account.accountId,
+          account.connection.networkId
+        )
+        const bytes = signedTx.encode()
+        sendTxnAsync = async () => {
+          await account.connection.provider.sendJsonRpc('broadcast_tx_async', [
+            Buffer.from(bytes).toString('base64'),
+          ])
+          console.log('TxHash', nearlib.utils.serialize.base_encode(txHash))
+        }
+        await sendTxnAsync()
+      }
+    } catch (e) {
+      // sleep to avoid socket hangout on retry too soon
+      await sleep(500)
+      continue
+    }
+
+    let result
+    for (let j = 0; j < RETRY_TX_STATUS; j++) {
+      try {
+        result = await account.connection.provider.txStatus(
+          txHash,
+          account.accountId
+        )
+        if (
+          result.status.SuccessValue !== undefined ||
+          result.status.Failure !== undefined
+        ) {
+          break
+        }
+      } catch (e) {
+        await sleep((j + 1) * 500)
+      }
+    }
+
+    if (result) {
+      const flatLogs = [
+        result.transaction_outcome,
+        ...result.receipts_outcome,
+      ].reduce((acc, it) => acc.concat(it.outcome.logs), [])
+      if (flatLogs && flatLogs != []) {
+        console.log(flatLogs)
+      }
+
+      if (result.status.SuccessValue !== undefined) {
+        return result
+      }
+
+      errorMsg = JSON.stringify(result.status.Failure)
+      if (errorMsg.includes('Transaction nonce')) {
+        // nonce incorrect, re-fetch nonce and retry
+        continue
+      } else {
+        // Indeed txn error, retry doesn't help
+        break
+      }
+    } else {
+      // Still no result after a long time, resubmit txn
+      resendLast = true
+      continue
+    }
+  }
+
+  throw new Error(errorMsg)
+}
+
 module.exports = {
   retry,
   sleep,
@@ -179,4 +303,6 @@ module.exports = {
   RobustWeb3,
   normalizeEthKey,
   promiseWithTimeout,
+  nearJsonContractFunctionCall,
+  signAndSendTransaction,
 }
